Refactor more actions
