+++ client/app/reader/DocumentList/DocumentListActions.js
@@ -2,8 +2,10 @@
 // setTagFilter, setCategoryFilter, selectCurrentPdfLocally, toggleDropdownFilterVisibility
 // setSearch, clearSearch, clearAllFilters
 
-import * as Constants from './constants';
-import { CATEGORIES } from './analytics';
+import * as Constants from '../constants';
+import ApiUtil from '../../util/ApiUtil';
+import { CATEGORIES, ENDPOINT_NAMES } from '../analytics';
+import { categoryFieldNameOfCategoryName } from '../utils';
 
 // Table header actions
 
@@ -67,6 +69,43 @@ export const toggleDropdownFilterVisibility = (filterName) => ({
     }
   }
 });
+
+export const toggleDocumentCategoryFail = (docId, categoryKey, categoryValueToRevertTo) => ({
+  type: Constants.TOGGLE_DOCUMENT_CATEGORY_FAIL,
+  payload: {
+    docId,
+    categoryKey,
+    categoryValueToRevertTo
+  }
+});
+
+export const handleCategoryToggle = (docId, categoryName, toggleState) => (dispatch) => {
+  const categoryKey = categoryFieldNameOfCategoryName(categoryName);
+
+  ApiUtil.patch(
+    `/document/${docId}`,
+    { data: { [categoryKey]: toggleState } },
+    ENDPOINT_NAMES.DOCUMENT
+  ).catch(() =>
+    dispatch(toggleDocumentCategoryFail(docId, categoryKey, !toggleState))
+  );
+
+  dispatch({
+    type: Constants.TOGGLE_DOCUMENT_CATEGORY,
+    payload: {
+      categoryKey,
+      toggleState,
+      docId
+    },
+    meta: {
+      analytics: {
+        category: CATEGORIES.VIEW_DOCUMENT_PAGE,
+        action: `${toggleState ? 'set' : 'unset'} document category`,
+        label: categoryName
+      }
+    }
+  });
+};
 
 // Tag filters
 
@@ -137,6 +176,34 @@ export const clearAllFilters = () => ({
     analytics: {
       category: CATEGORIES.CLAIMS_FOLDER_PAGE,
       action: 'clear-all-filters'
+    }
+  }
+});
+
+export const setViewingDocumentsOrComments = (documentsOrComments) => ({
+  type: Constants.SET_VIEWING_DOCUMENTS_OR_COMMENTS,
+  payload: {
+    documentsOrComments
+  },
+  meta: {
+    analytics: {
+      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
+      action: 'set-viewing-documents-or-comments',
+      label: documentsOrComments
+    }
+  }
+});
+
+export const handleToggleCommentOpened = (docId) => ({
+  type: Constants.TOGGLE_COMMENT_LIST,
+  payload: {
+    docId
+  },
+  meta: {
+    analytics: {
+      category: CATEGORIES.CLAIMS_FOLDER_PAGE,
+      action: 'toggle-comment-list',
+      label: (nextState) => nextState.readerReducer.documents[docId].listComments ? 'open' : 'close'
     }
   }
 });
